Trait buildout for testing param parsing trait implementers.
