+++ tests/S3/Crypto/S3EncryptionClientTest.php
@@ -12,6 +12,7 @@ use Aws\Crypto\MetadataEnvelope;
 use Aws\S3\S3Client;
 use Aws\S3\Crypto\HeadersMetadataStrategy;
 use Aws\S3\Crypto\InstructionFileMetadataStrategy;
+use Aws\Test\Crypto\UsesCryptoParamsTrait;
 use Aws\Test\UsesServiceTrait;
 use Aws\Test\Crypto\UsesMetadataEnvelopeTrait;
 use GuzzleHttp\Promise\FulfilledPromise;
@@ -21,9 +22,9 @@ use Psr\Http\Message\RequestInterface;
 
 class S3EncryptionClientTest extends \PHPUnit_Framework_TestCase
 {
-    use UsesServiceTrait, UsesMetadataEnvelopeTrait;
+    use UsesServiceTrait, UsesMetadataEnvelopeTrait, UsesCryptoParamsTrait;
 
-    private function getS3Client()
+    protected function getS3Client()
     {
         static $client = null;
 
@@ -33,7 +34,7 @@ class S3EncryptionClientTest extends \PHPUnit_Framework_TestCase
         return $client;
     }
 
-    private function getKmsClient()
+    protected function getKmsClient()
     {
         static $client = null;
 
@@ -41,33 +42,6 @@ class S3EncryptionClientTest extends \PHPUnit_Framework_TestCase
             $client = $this->getTestClient('Kms');
         }
         return $client;
-    }
-
-    public function getValidMaterialsProviders()
-    {
-        $kms = $this->getKmsClient();
-        $keyId = '11111111-2222-3333-4444-555555555555';
-        return [
-            [
-                new KmsMaterialsProvider($kms, $keyId),
-                false
-            ]
-        ];
-    }
-
-    public function getInvalidMaterialsProviders()
-    {
-
-        return [
-            [
-                new \stdClass(),
-                [
-                    'InvalidArgumentException',
-                    'An instance of MaterialsProvider'
-                    . ' must be passed in the "MaterialsProvider" field.'
-                ]
-            ]
-        ];
     }
 
     private function setupProvidedExpectedException($exception)
@@ -137,62 +111,6 @@ class S3EncryptionClientTest extends \PHPUnit_Framework_TestCase
                 'Cipher' => 'cbc'
             ]
         ]);
-    }
-
-    public function getValidMetadataStrategies()
-    {
-        $s3 = $this->getS3Client();
-        return [
-            [
-                null,
-                false,
-                1
-            ],
-            [
-                HeadersMetadataStrategy::class,
-                false,
-                1
-            ],
-            [
-                InstructionFileMetadataStrategy::class,
-                false,
-                2
-            ],
-            [
-                new HeadersMetadataStrategy(),
-                false,
-                1
-            ],
-            [
-                new InstructionFileMetadataStrategy($s3),
-                false,
-                2
-            ]
-        ];
-    }
-
-    public function getInvalidMetadataStrategies()
-    {
-        return [
-            [
-                new \stdClass(),
-                [
-                    'InvalidArgumentException',
-                    'The metadata strategy that'
-                    . ' was passed to "MetadataStrategy" was unrecognized.'
-                ]
-
-            ],
-            [
-                'Not Predefined Strategy',
-                [
-                    'InvalidArgumentException',
-                    'Could not match the'
-                    . ' specified string in "MetadataStrategy" to a'
-                    . ' predefined strategy.'
-                ]
-            ]
-        ];
     }
 
     /**
@@ -324,29 +242,6 @@ class S3EncryptionClientTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue($this->mockQueueEmpty());
     }
 
-    public function getCiphers()
-    {
-        return [
-            [
-                'cbc'
-            ],
-            [
-                'gcm',
-                null,
-                function () {
-                    return version_compare(PHP_VERSION, '7.1', '<');
-                }
-            ],
-            [
-                'unsupported',
-                [
-                    'InvalidArgumentException',
-                    'The cipher requested is not supported by the SDK.'
-                ]
-            ]
-        ];
-    }
-
     /**
      * @dataProvider getCiphers
      */
@@ -387,39 +282,6 @@ class S3EncryptionClientTest extends \PHPUnit_Framework_TestCase
                 'Cipher' => $cipher
             ]
         ]);
-    }
-
-    public function getKeySizes()
-    {
-        return [
-            [
-                128,
-                []
-            ],
-            [
-                192,
-                []
-            ],
-            [
-                256,
-                []
-            ],
-            [
-                'gcm',
-                [
-                    'InvalidArgumentException',
-                    'The cipher "KeySize" must be an integer.'
-                ]
-            ],
-            [
-                512,
-                [
-                    'InvalidArgumentException',
-                    'The cipher "KeySize" requested'
-                    . ' is not supported by AES (128, 192, or 256).'
-                ]
-            ]
-        ];
     }
 
     /**
